+++ sdk/python/packages/flet-core/src/flet_core/__init__.py
@@ -16,7 +16,7 @@ from flet_core import (
 from flet_core.alert_dialog import AlertDialog
 from flet_core.alignment import Alignment
 from flet_core.animated_switcher import AnimatedSwitcher
-from flet_core.animation import Animation
+from flet_core.animation import Animation, AnimationCurve
 from flet_core.app_bar import AppBar
 from flet_core.audio import Audio
 from flet_core.banner import Banner
@@ -187,7 +187,6 @@ from flet_core.types import (
     WEB_BROWSER,
     AppView,
     AnimatedSwitcherTransition,
-    AnimationCurve,
     BlendMode,
     BoxShape,
     ClipBehavior,
