+++ sdk/python/packages/flet-core/src/flet_core/container.py
@@ -27,9 +27,8 @@ from flet_core.types import (
     RotateValue,
     ScaleValue,
     ThemeMode,
-    get_valid_enum,
-    get_non_default_value,
 )
+from flet_core.utils import get_valid_enum, get_non_default_value
 
 
 _BlendModeDefault = BlendMode.MODULATE
