+++ sdk/python/packages/flet-core/src/flet_core/row.py
@@ -13,9 +13,8 @@ from flet_core.types import (
     RotateValue,
     ScaleValue,
     ScrollMode,
-    get_valid_enum,
-    get_non_default_value,
 )
+from flet_core.utils import get_valid_enum, get_non_default_value
 
 
 _CrossAxisAlignmentDefault = CrossAxisAlignment.START
