+++ sdk/python/packages/flet-core/src/flet_core/stack.py
@@ -10,9 +10,8 @@ from flet_core.types import (
     ResponsiveNumber,
     RotateValue,
     ScaleValue,
-    get_valid_enum,
-    get_non_default_value,
 )
+from flet_core.utils import get_valid_enum, get_non_default_value
 
 
 _ClipBehaviorDefault = ClipBehavior.HARD_EDGE
