+++ sdk/python/packages/flet-core/src/flet_core/textfield.py
@@ -15,9 +15,9 @@ from flet_core.types import (
     RotateValue,
     ScaleValue,
     TextAlign,
-    get_valid_enum,
-    get_non_default_value,
 )
+from flet_core.utils import get_valid_enum, get_non_default_value
+
 
 _TextAlignDefault = TextAlign.LEFT
